+++ lib/pgsync.rb
@@ -16,9 +16,10 @@ require "pgsync/utils"
 require "pgsync/client"
 require "pgsync/data_source"
 require "pgsync/init"
+require "pgsync/schema_sync"
 require "pgsync/sync"
-require "pgsync/table_list"
 require "pgsync/table_sync"
+require "pgsync/task_resolver"
 require "pgsync/version"
 
 module PgSync