+++ lib/pgsync/sync.rb
@@ -2,21 +2,26 @@ module PgSync
   class Sync
     include Utils
 
-    def perform(options)
-      args = options.arguments
-      opts = options.to_hash
-      @options = opts
+    def initialize(arguments, options)
+      @arguments = arguments
+      @options = options
+    end
 
-      # merge config
-      [:to, :from, :to_safe, :exclude, :schemas].each do |opt|
+    def perform
+      args = @arguments
+      opts = @options
+
+      # only resolve commands from config, not CLI arguments
+      [:to, :from].each do |opt|
+        opts[opt] ||= resolve_source(config[opt.to_s])
+      end
+
+      # merge other config
+      [:to_safe, :exclude, :schemas].each do |opt|
         opts[opt] ||= config[opt.to_s]
       end
 
-      # TODO remove deprecations in 0.6.0
-      map_deprecations(args, opts)
-
-      # start
-      start_time = Time.now
+      started_at = Time.now
 
       if args.size > 2
         raise Error, "Usage:\n    pgsync [options]"
@@ -29,20 +34,32 @@ module PgSync
         raise Error, "Danger! Add `to_safe: true` to `.pgsync.yml` if the destination is not localhost or 127.0.0.1"
       end
 
+      if (opts[:preserve] || opts[:overwrite]) && destination.server_version_num < 90500
+        raise Error, "Postgres 9.5+ is required for --preserve and --overwrite"
+      end
+
       print_description("From", source)
       print_description("To", destination)
 
-      tables = TableList.new(args, opts, source, config).tables
-      confirm_tables_exist(source, tables, "source")
+      tasks = TaskResolver.new(args, opts, source, config).tasks
+      tasks.map! do |task|
+        TableSync.new(source: source, destination: destination, config: config, table: task[:table], opts: opts.merge(sql: task[:sql]))
+      end
+
+      if opts[:in_batches] && tasks.size > 1
+        raise Error, "Cannot use --in-batches with multiple tables"
+      end
+
+      confirm_tables_exist(source, tasks, "source")
 
       if opts[:list]
-        confirm_tables_exist(destination, tables, "destination")
+        confirm_tables_exist(destination, tasks, "destination")
 
         list_items =
           if args[0] == "groups"
             (config["groups"] || {}).keys
           else
-            tables.keys
+            tasks.map(&:table)
           end
 
         pretty_list list_items
@@ -53,66 +70,71 @@ module PgSync
           end
 
           log "* Dumping schema"
-          schema_tables = tables if !opts[:all_schemas] || opts[:tables] || opts[:groups] || args[0] || opts[:exclude]
-          sync_schema(source, destination, schema_tables)
+          schema_tasks =
+            if !opts[:all_schemas] || opts[:tables] || opts[:groups] || args[0] || opts[:exclude]
+              tasks
+            end
+          SchemaSync.new(source: source, destination: destination, tasks: schema_tasks).perform
         end
 
         unless opts[:schema_only]
-          confirm_tables_exist(destination, tables, "destination")
+          confirm_tables_exist(destination, tasks, "destination")
 
-          in_parallel(tables, first_schema: source.search_path.find { |sp| sp != "pg_catalog" }) do |table, table_opts, source, destination|
-            TableSync.new(source: source, destination: destination).sync(config, table, opts.merge(table_opts))
+          # TODO only query specific tables
+          # TODO add sequences, primary keys, etc
+          source_columns = columns(source)
+          destination_columns = columns(destination)
+
+          tasks.each do |task|
+            task.from_columns = source_columns[task.table] || []
+            task.to_columns = destination_columns[task.table] || []
+          end
+
+          # show notes before we start
+          tasks.each do |task|
+            task.notes.each do |note|
+              warning "#{display_task(task)}: #{note}"
+            end
+          end
+
+          # don't sync tables with no shared fields
+          # we show a warning message above
+          tasks.reject! { |task| task.shared_fields.empty? }
+
+          in_parallel(tasks) do |task|
+            task.sync
           end
         end
 
-        log_completed(start_time)
+        log_completed(started_at)
       end
     end
 
-    def confirm_tables_exist(data_source, tables, description)
-      tables.keys.each do |table|
+    def columns(data_source)
+      query = <<~SQL
+        SELECT
+          table_schema AS schema,
+          table_name AS table,
+          column_name AS column,
+          data_type AS type
+        FROM
+          information_schema.columns
+        ORDER BY 1, 2, 3
+      SQL
+      data_source.execute(query).group_by { |r| [r["schema"], r["table"]] }.map do |k, v|
+        [k.join("."), v.map { |r| {name: r["column"], type: r["type"]} }]
+      end.to_h
+    end
+
+    def first_schema
+      @first_schema ||= source.search_path.find { |sp| sp != "pg_catalog" }
+    end
+
+    def confirm_tables_exist(data_source, tasks, description)
+      tasks.map(&:table).each do |table|
         unless data_source.table_exists?(table)
-          raise Error, "Table does not exist in #{description}: #{table}"
+          raise Error, "Table not found in #{description}: #{table}"
         end
-      end
-    end
-
-    def map_deprecations(args, opts)
-      command = args[0]
-
-      case command
-      when "schema"
-        args.shift
-        opts[:schema_only] = true
-        deprecated "Use `psync --schema-only` instead"
-      when "tables"
-        args.shift
-        opts[:tables] = args.shift
-        deprecated "Use `pgsync #{opts[:tables]}` instead"
-      when "groups"
-        args.shift
-        opts[:groups] = args.shift
-        deprecated "Use `pgsync #{opts[:groups]}` instead"
-      end
-
-      if opts[:where]
-        opts[:sql] ||= String.new
-        opts[:sql] << " WHERE #{opts[:where]}"
-        deprecated "Use `\"WHERE #{opts[:where]}\"` instead"
-      end
-
-      if opts[:limit]
-        opts[:sql] ||= String.new
-        opts[:sql] << " LIMIT #{opts[:limit]}"
-        deprecated "Use `\"LIMIT #{opts[:limit]}\"` instead"
-      end
-    end
-
-    def sync_schema(source, destination, tables = nil)
-      dump_command = source.dump_command(tables)
-      restore_command = destination.restore_command
-      unless system("#{dump_command} | #{restore_command}")
-        raise Error, "Schema sync returned non-zero exit code"
       end
     end
 
@@ -124,6 +146,8 @@ module PgSync
             YAML.load_file(file) || {}
           rescue Psych::SyntaxError => e
             raise Error, e.message
+          rescue Errno::ENOENT
+            raise Error, "Config file not found: #{file}"
           end
         else
           {}
@@ -136,15 +160,14 @@ module PgSync
       log "#{prefix}: #{source.dbname}#{location}"
     end
 
-    def in_parallel(tables, first_schema:, &block)
-      spinners = TTY::Spinner::Multi.new(format: :dots, output: output)
-      item_spinners = {}
+    def in_parallel(tasks, &block)
+      failed_tables = []
 
-      start = lambda do |item, i|
-        table, opts = item
-        message = String.new(":spinner ")
-        message << table.sub("#{first_schema}.", "")
-        message << " #{opts[:sql]}" if opts[:sql]
+      spinners = TTY::Spinner::Multi.new(format: :dots, output: output)
+      task_spinners = {}
+
+      start = lambda do |task, i|
+        message = ":spinner #{display_item(task)}"
         spinner = spinners.register(message)
         if @options[:in_batches]
           # log instead of spin for non-tty
@@ -152,27 +175,25 @@ module PgSync
         else
           spinner.auto_spin
         end
-        item_spinners[item] = spinner
+        task_spinners[task] = spinner
       end
 
-      failed_tables = []
-
-      finish = lambda do |item, i, result|
-        spinner = item_spinners[item]
-        table_name = item.first.sub("#{first_schema}.", "")
+      finish = lambda do |task, i, result|
+        spinner = task_spinners[task]
+        result_message = display_result(result)
 
         if result[:status] == "success"
-          spinner.success(display_message(result))
+          spinner.success(result_message)
         else
           # TODO add option to fail fast
-          spinner.error(display_message(result))
-          failed_tables << table_name
+          spinner.error(result_message)
+          failed_tables << display_task(task)
           fail_sync(failed_tables) if @options[:fail_fast]
         end
 
         unless spinner.send(:tty?)
           status = result[:status] == "success" ? "✔" : "✖"
-          log [status, table_name, item.last[:sql], display_message(result)].compact.join(" ")
+          log [status, display_item(task), result_message].join(" ")
         end
       end
 
@@ -180,7 +201,7 @@ module PgSync
 
       jobs = @options[:jobs]
       if @options[:debug] || @options[:in_batches] || @options[:defer_constraints]
-        warn "--jobs ignored" if jobs
+        warning "--jobs ignored" if jobs
         jobs = 0
       end
 
@@ -194,17 +215,11 @@ module PgSync
         # could try to use `raise Parallel::Kill` to fail faster with --fail-fast
         # see `fast_faster` branch
         # however, need to make sure connections are cleaned up properly
-        Parallel.each(tables, **options) do |table, table_opts|
-          # must reconnect for new thread or process
-          # TODO only reconnect first time
-          unless options[:in_processes] == 0
-            source.reconnect
-            destination.reconnect
-          end
+        Parallel.each(tasks, **options) do |task|
+          source.reconnect_if_needed
+          destination.reconnect_if_needed
 
-          # TODO warn if there are non-deferrable constraints on the table
-
-          yield table, table_opts, source, destination
+          yield task
         end
       end
 
@@ -231,7 +246,18 @@ module PgSync
       raise Error, "Sync failed for #{failed_tables.size} table#{failed_tables.size == 1 ? nil : "s"}: #{failed_tables.join(", ")}"
     end
 
-    def display_message(result)
+    def display_task(task)
+      task.table.sub("#{first_schema}.", "")
+    end
+
+    def display_item(item)
+      messages = []
+      messages << display_task(item)
+      messages << item.opts[:sql] if item.opts[:sql]
+      messages.join(" ")
+    end
+
+    def display_result(result)
       messages = []
       messages << "- #{result[:time]}s" if result[:time]
       messages << "(#{result[:message].lines.first.to_s.strip})" if result[:message]
@@ -244,12 +270,8 @@ module PgSync
       end
     end
 
-    def deprecated(message)
-      log colorize("[DEPRECATED] #{message}", :yellow)
-    end
-
-    def log_completed(start_time)
-      time = Time.now - start_time
+    def log_completed(started_at)
+      time = Time.now - started_at
       message = "Completed in #{time.round(1)}s"
       log colorize(message, :green)
     end
@@ -270,6 +292,21 @@ module PgSync
       ds = DataSource.new(url)
       ObjectSpace.define_finalizer(self, self.class.finalize(ds))
       ds
+    end
+
+    def resolve_source(source)
+      if source
+        source = source.dup
+        source.gsub!(/\$\([^)]+\)/) do |m|
+          command = m[2..-2]
+          result = `#{command}`.chomp
+          unless $?.success?
+            raise Error, "Command exited with non-zero status:\n#{command}"
+          end
+          result
+        end
+      end
+      source
     end
 
     def self.finalize(ds)