+++ lib/pgsync/utils.rb
@@ -18,16 +18,31 @@ module PgSync
       end
     end
 
+    def warning(message)
+      log colorize(message, :yellow)
+    end
+
+    def deprecated(message)
+      warning "[DEPRECATED] #{message}"
+    end
+
     def output
       $stderr
     end
 
     def config_file
-      search_tree(db_config_file(@options[:db]) || @options[:config] || ".pgsync.yml")
+      if @options[:config]
+        @options[:config]
+      elsif @options[:db]
+        file = db_config_file(@options[:db])
+        search_tree(file) || file
+      else
+        search_tree(".pgsync.yml")
+      end
     end
 
     def db_config_file(db)
-      ".pgsync-#{db}.yml" if db
+      ".pgsync-#{db}.yml"
     end
 
     def search_tree(file)