+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.6.0 (unreleased)
+
+- Added message for different column types
+
+Breaking
+
+- Dropped support for Postgres < 9.5
+- Removed deprecated options
+- Raise an error when `--config` or `--db` option provided and config not found
+
 ## 0.5.5 (2020-05-13)
 
 - Added `--jobs` option